chore: extract tz functions
