+++ src/tz/index.ts
@@ -1,1 +1,2 @@
 export * from './api';
+export * from './utils';
